reorganize types files.
