+++ generators/base/application.d.mts
