+++ generators/bootstrap-application-base/generator.mts
@@ -25,7 +25,7 @@ import { prepareRelationshipForTemplates } from '../../utils/relationship.mjs';
 import { stringify } from '../../utils/index.mjs';
 import { createUserEntity } from './utils.mjs';
 import { DOCKER_DIR, NODE_VERSION } from '../generator-constants.mjs';
-import type { CommonClientServerApplication } from './types.js';
+import type { CommonClientServerApplication } from '../base-application/types.mjs';
 import { GENERATOR_BOOTSTRAP } from '../generator-list.mjs';
 import { addFakerToEntity } from './faker.mjs';
