--- generators/bootstrap-application/types.d.ts
@@ -1,4 +1,0 @@
-import type { ClientApplication } from '../bootstrap-application-client/types';
-import type { SpringBootApplication } from '../bootstrap-application-server/types';
-
-export type ClientServerApplication = ClientApplication & SpringBootApplication;
