+++ generators/common/types.d.mts
@@ -1,0 +1,4 @@
+import type { ClientApplication } from '../client/types.mjs';
+import type { SpringBootApplication } from '../server/types.mjs';
+
+export type ClientServerApplication = ClientApplication & SpringBootApplication;
