+++ generators/export-jdl/generator.mts
@@ -24,7 +24,7 @@ import statistics from '../statistics.cjs';
 import { GENERATOR_EXPORT_JDL } from '../generator-list.mjs';
 import applicationOptions from '../../jdl/jhipster/application-options.js';
 import JSONToJDLConverter from '../../jdl/converters/json-to-jdl-converter.js';
-import type { JHipsterGeneratorOptions, JHipsterGeneratorFeatures } from '../base/api.cjs';
+import type { JHipsterGeneratorOptions, JHipsterGeneratorFeatures } from '../base/api.mjs';
 
 const { OptionNames } = applicationOptions;
